requirements.txt
openai
azure-core
azure-ai-formrecognizer
azure-ai-textanalytics
python-dotenv